feat: landing navigability through page and link. feat: styling
added font to index.html
added styling to page
added navigation through page

missing: information for sections
missing: dropdown navigation for small devices
