+++ src/components/footer/Footer.tsx
@@ -1,8 +1,8 @@
-import {FC} from 'react'
+import { FC } from 'react'
 import footerLogo from '../../assets/icons/footer-logo.png'
 import { FaInstagram, FaFacebook, FaTwitter } from 'react-icons/fa'
 
-const Footer:FC = () => {
+const Footer: FC = () => {
     return (
         <footer className="bg-gray-900 text-white py-10 px-4">
             {/* Top Section */}
@@ -12,9 +12,9 @@ const Footer:FC = () => {
                     <img src={footerLogo} alt="Logo" className="mb-5 w-36" />
                     <ul className="flex flex-col md:flex-row gap-4">
                         <li><a href="#home" className="hover:text-primary">Inicio</a></li>
-                        <li><a href="#services" className="hover:text-primary">Nosotros</a></li>
-                        <li><a href="#about" className="hover:text-primary">Cómo funciona</a></li>
-                        <li><a href="#contact" className="hover:text-primary">Servicios</a></li>
+                        <li><a href="#us" className="hover:text-primary">Nosotros</a></li>
+                        <li><a href="#how-it-works" className="hover:text-primary">Cómo funciona</a></li>
+                        <li><a href="#services" className="hover:text-primary">Servicios</a></li>
                     </ul>
                 </div>
