+++ tailwind.config.js
@@ -7,13 +7,17 @@ module.exports = {
   theme: {
     fontFamily: {
       'sans': ['Roboto', 'sans-serif'],
-      'mono': ['Fira Code', 'monospace'],
+      'mono': ['"Fira Code"', 'monospace'],
+      'landing': ['Raleway','sans-serif']
     },
     extend: {
+      backgroundImage: {
+        'landing-image': "url('/src/assets/landing-bg.svg')",
+      },
       colors: {
         'black': '#162D3A',
         'landing-primary': '#F5F8FF',
-        'landing-secondary':'#252432',
+        'landing-secondary': '#252432',
         'background-input-login': '#F7FBFF',
         'strike-input-login': '#D4D7E3',
         'blue': '#1E4AE9',
